+++ content/event/hpedev-hack-shack.md
@@ -12,5 +12,3 @@ width: large
 ## The HPE DEV Hack Shack
 A continuing virtual event
 The HPE Developer Community presents the HPE DEV Hack Shack, a free, online community gathering place that’s always open. Gather with other developers, data scientists, and technologists to get hands-on experience with popular open source and HPE technologies in our Workshops-on-Demand. Learn more about these technologies in our video replays and HPE Ezmeral Learn On Demand library. Pose questions to our Slack channel, sign up for our Newsletter, or connect to our HPE DEV portal, where you can access numerous tutorials and GitHub repositories. Or simply spend some time relaxing and playing games. You can access it all here, from wherever you are.
-
-[Learn more](https://hackshack.hpedev.io/)
