+++ content/event/hpedev-much-learn.md
@@ -12,6 +12,4 @@ width: large
 ## HPE DEV Munch & Learn series
 What's a data fabric and how does it work?
 January 27, 2021
-Run by the HPE Developer Community, the HPE DEV Munch & Learn monthly technology talks feature industry technology experts who offer valuable insights into today’s most popular technologies. Take advantage of this free 60-minute session opportunity to listen to and engage with these leading technologists on a variety of subjects. Each month we will be presenting a different topic. For January 2021, meet with Ted Dunning and learn about the value of a data fabric and how it works..
-
-[Register](https://hpe.zoom.us/meeting/register/tJwrc-ygrjItG90Z60OnGVisrppOzsT_N0fd)
+Run by the HPE Developer Community, the HPE DEV Munch & Learn monthly technology talks feature industry technology experts who offer valuable insights into today’s most popular technologies. Take advantage of this free 60-minute session opportunity to listen to and engage with these leading technologists on a variety of subjects. Each month we will be presenting a different topic. For January 2021, meet with Ted Dunning and learn about the value of a data fabric and how it works.
