+++ content/event/kube-con-europe-2021.md
@@ -12,5 +12,3 @@ width: large
 ## KubeCon Europe 2021
 May 04-07, 2021 Europe
 The Cloud Native Computing Foundation’s flagship conference gathers adopters and technologists from leading open source and cloud native communities virtually from May 4 – 7, 2021. Join containerd, CoreDNS, Envoy, Fluentd, Harbor, Helm, Jaeger, Kubernetes, Prometheus, Rook, TiKV, TUF, Vitess, Argo, CloudEvents, CNI, Contour, Cortex, CRI-O, Dragonfly, etcd, Falco, gRPC, KubeEdge, Linkerd, NATS, Notary, Open Policy Agent, OpenTracing, Operator Framework, SPIFFE, SPIRE, and Thanos as the community gathers for four days to further the education and advancement of cloud native computing.
-
-[View Event Page](https://events.linuxfoundation.org/kubecon-cloudnativecon-europe/)
