+++ content/event/kube-con-na-2020.md
@@ -12,5 +12,3 @@ width: large
 ## KubeCon NA 2020
 November 17-20, 2020 NA
 The Cloud Native Computing Foundation’s flagship conference gathers adopters and technologists from leading open source and cloud native communities virtually from November 17 – 20, 2020. Join Kubernetes, Prometheus, Envoy, CoreDNS, containerd, Fluentd, Jaeger, Vitess, TUF, OpenTracing, gRPC, CNI, Notary, NATS, Linkerd, Helm, Rook, Harbor, etcd, Open Policy Agent, CRI-O, TiKV, CloudEvents, Falco, Argo, Dragonfly, SPIFFE and SPIRE as the community gathers for four days to further the education and advancement of cloud native computing.
-
-[View Event Page](https://events.linuxfoundation.org/events/kubecon-cloudnativecon-north-america-2019/)
