+++ content/event/technology-and-solutions-summit-2020-paris.md
@@ -12,5 +12,3 @@ width: large
 ## TSS Paris 2020
 March 9-13, 2020 Paris
 HPE TSS is Hewlett Packard Enterprise's largest, annual and most comprehensive technical and solutions knowledge-transfer event. Since its inception in 2006, HPE TSS has established its reputation as a renowned training initiative. Presales Consultants and Solutions Architects from HPE and partner communities in Europe, Middle East and Africa benefit from direct access to HPE technology experts, engineers and Chief Technologists. You will receive up to the minute training around product developments, roadmap updates, strategic insights and be able to practice and apply your knowledge under non-disclosure.
-
-[View Event Page](https://h41382.www4.hpe.com/tss/)
